vm: fix invocation without architecture
When no architecture is specified, the x86_64 arch is the default,
but its parser was not invoked and thus the defaults of that sub-
parser have not been set.
Convert the Namespace object to a normal dict and use .get in the
x86_64 case to ensure we have proper defaults in that case too.